Revert "Allow BCCSP config to be set using env var (#1900)"
This reverts commit f2e9e6e7b307a25dc4738e97a1ffb08daea4f6dc.
which pushed the pkcs11 code outside of the pkcs11 build tag.
This prevents cross-compilation of Fabric due to the fact the
miekg/pkcs11 package contains CGO which requires a cross-compiler
for each platform you are cross compiling for.

We need to find a better approach to solving this problem
that doesn't pull the code out of the build tag.
