+++ bccsp/factory/opts.go
@@ -1,19 +1,20 @@
 /*
-Copyright IBM Corp. All Rights Reserved.
+Copyright IBM Corp. 2016 All Rights Reserved.
 
-SPDX-License-Identifier: Apache-2.0
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+		 http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
 */
 
 package factory
-
-import "github.com/hyperledger/fabric/bccsp/pkcs11"
-
-// FactoryOpts holds configuration information used to initialize factory implementations
-type FactoryOpts struct {
-	ProviderName string             `mapstructure:"default" json:"default" yaml:"Default"`
-	SwOpts       *SwOpts            `mapstructure:"SW,omitempty" json:"SW,omitempty" yaml:"SW,omitempty"`
-	Pkcs11Opts   *pkcs11.PKCS11Opts `mapstructure:"PKCS11,omitempty" json:"PKCS11,omitempty" yaml:"PKCS11"`
-}
 
 // GetDefaultOpts offers a default implementation for Opts
 // returns a new instance every time
@@ -24,10 +25,6 @@ func GetDefaultOpts() *FactoryOpts {
 			HashFamily: "SHA2",
 			SecLevel:   256,
 			Ephemeral:  true,
-		},
-		Pkcs11Opts: &pkcs11.PKCS11Opts{
-			HashFamily: "SHA2",
-			SecLevel:   256,
 		},
 	}
 }
