+++ common/viperutil/config_util.go
@@ -306,7 +306,7 @@ func bccspHook(f reflect.Type, t reflect.Type, data interface{}) (interface{}, e
 
 	err := mapstructure.WeakDecode(data, config)
 	if err != nil {
-		return nil, errors.Wrap(err, "could not decode bccsp type")
+		return nil, errors.Wrap(err, "could not decode bcssp type")
 	}
 
 	return config, nil
