+++ internal/peer/common/common.go
@@ -12,7 +12,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"strconv"
 	"strings"
 	"time"
 
@@ -134,7 +133,6 @@ func InitCrypto(mspMgrConfigDir, localMSPID, localMSPType string) error {
 
 	// Init the BCCSP
 	SetBCCSPKeystorePath()
-
 	bccspConfig := factory.GetDefaultOpts()
 	if config := viper.Get("peer.BCCSP"); config != nil {
 		err = mapstructure.WeakDecode(config, bccspConfig)
@@ -143,48 +141,9 @@ func InitCrypto(mspMgrConfigDir, localMSPID, localMSPType string) error {
 		}
 	}
 
-	if err := SetBCCSPConfigOverrides(bccspConfig); err != nil {
-		return err
-	}
-
 	err = mspmgmt.LoadLocalMspWithType(mspMgrConfigDir, bccspConfig, localMSPID, localMSPType)
 	if err != nil {
 		return errors.WithMessagef(err, "error when setting up MSP of type %s from directory %s", localMSPType, mspMgrConfigDir)
-	}
-
-	return nil
-}
-
-// Overrides BCCSP config values when corresponding environment variables
-// are set.
-func SetBCCSPConfigOverrides(bccspConfig *factory.FactoryOpts) error {
-	if pkcs11Default, exist := os.LookupEnv("CORE_PEER_BCCSP_DEFAULT"); exist {
-		bccspConfig.ProviderName = pkcs11Default
-	}
-
-	// PKCS11 Overrides
-	if pkcs11Hash, exist := os.LookupEnv("CORE_PEER_BCCSP_PKCS11_HASH"); exist {
-		bccspConfig.Pkcs11Opts.HashFamily = pkcs11Hash
-	}
-
-	if pkcs11Security, exist := os.LookupEnv("CORE_PEER_BCCSP_PKCS11_SECURITY"); exist {
-		pkcs11Sec, err := strconv.Atoi(pkcs11Security)
-		if err != nil {
-			return errors.Errorf("CORE_PEER_BCCSP_PKCS11_SECURITY set to non-integer value: %s", pkcs11Security)
-		}
-		bccspConfig.Pkcs11Opts.SecLevel = pkcs11Sec
-	}
-
-	if pkcs11Library, exist := os.LookupEnv("CORE_PEER_BCCSP_PKCS11_LIBRARY"); exist {
-		bccspConfig.Pkcs11Opts.Library = pkcs11Library
-	}
-
-	if pkcs11Label, exist := os.LookupEnv("CORE_PEER_BCCSP_PKCS11_LABEL"); exist {
-		bccspConfig.Pkcs11Opts.Label = pkcs11Label
-	}
-
-	if pkcs11Pin, exist := os.LookupEnv("CORE_PEER_BCCSP_PKCS11_PIN"); exist {
-		bccspConfig.Pkcs11Opts.Pin = pkcs11Pin
 	}
 
 	return nil
